removed unused code to reflect changes
